+++ lib/ood_core/job/adapters/lsf/helper.rb
@@ -88,6 +88,7 @@ class OodCore::Job::Adapters::Lsf::Helper
     args += (script.rerunnable ? ["-r"] : ["-rn"]) unless script.rerunnable.nil?
     args += ["-b", script.start_time.localtime.strftime("%Y:%m:%d:%H:%M")] unless script.start_time.nil?
     args += ["-W", (script.wall_time / 60).to_i] unless script.wall_time.nil?
+    args += ["-L", script.shell_path.to_s] unless script.shell_path.nil?
 
     # input and output files
     args += ["-i", script.input_path] unless script.input_path.nil?